# platform/backend/src/app/DTOs/CreatePatientDto.cs
using System;

namespace Hospital.Application.DTOs
{
    /// <summary>
    /// DTO for creating a new Patient.
    /// Only contains fields that can be set during creation.
    /// Excludes: Id (auto-generated), CreatedAt/UpdatedAt (auto-set), IsActive (defaults to true)
    /// </summary>
    public class CreatePatientDto
    {
        public string FirstName { get; set; } = string.Empty;
        public string LastName { get; set; } = string.Empty;
        public DateTime DateOfBirth { get; set; }
        public string Gender { get; set; } = string.Empty;
        public string? Email { get; set; }
        public string PhoneNumber { get; set; } = string.Empty;
        public string? Address { get; set; }
        public string? City { get; set; }
        public string? State { get; set; }
        public string? ZipCode { get; set; }
        public string? BloodGroup { get; set; }
        public string? EmergencyContactName { get; set; }
        public string? EmergencyContactPhone { get; set; }
        public string? MedicalHistory { get; set; }
        public string? Allergies { get; set; }
    }
}