# platform/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Hospital Patient Management API
  description: API for managing patient information in a hospital system
  version: 1.0.0
  contact:
    name: Hospital IT Department
    email: it@hospital.com

servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.hospital.com/api
    description: Production server

tags:
  - name: Patients
    description: Patient management operations

paths:
  /patients:
    get:
      tags:
        - Patients
      summary: Get all patients
      description: Retrieve a list of all active patients
      operationId: getAllPatients
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientDto'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Patients
      summary: Create a new patient
      description: Register a new patient in the system
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientDto'
      responses:
        '201':
          description: Patient created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDto'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal server error

  /patients/{id}:
    get:
      tags:
        - Patients
      summary: Get patient by ID
      description: Retrieve a specific patient by their unique identifier
      operationId: getPatientById
      parameters:
        - name: id
          in: path
          required: true
          description: Patient unique identifier
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDto'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Patients
      summary: Update patient
      description: Update an existing patient's information
      operationId: updatePatient
      parameters:
        - name: id
          in: path
          required: true
          description: Patient unique identifier
          schema:
            type: integer
            format: int32
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientDto'
      responses:
        '200':
          description: Patient updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDto'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Patient not found

    delete:
      tags:
        - Patients
      summary: Delete patient
      description: Soft delete a patient (marks as inactive)
      operationId: deletePatient
      parameters:
        - name: id
          in: path
          required: true
          description: Patient unique identifier
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Patient deleted successfully
        '404':
          description: Patient not found

  /patients/search:
    get:
      tags:
        - Patients
      summary: Search patients
      description: Search patients by name, email, or phone number
      operationId: searchPatients
      parameters:
        - name: query
          in: query
          required: true
          description: Search term
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientDto'

components:
  schemas:
    PatientDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: Unique identifier for the patient
        firstName:
          type: string
          description: Patient's first name
          example: John
        lastName:
          type: string
          description: Patient's last name
          example: Doe
        fullName:
          type: string
          description: Patient's full name
          example: John Doe
        dateOfBirth:
          type: string
          format: date
          description: Patient's date of birth
          example: '1985-03-15'
        age:
          type: integer
          description: Patient's calculated age
          example: 38
        gender:
          type: string
          enum: [Male, Female, Other]
          description: Patient's gender
          example: Male
        email:
          type: string
          format: email
          nullable: true
          description: Patient's email address
          example: john.doe@email.com
        phoneNumber:
          type: string
          description: Patient's phone number
          example: '+1-555-0101'
        address:
          type: string
          nullable: true
          description: Patient's street address
          example: 123 Main Street
        city:
          type: string
          nullable: true
          description: Patient's city
          example: New York
        state:
          type: string
          nullable: true
          description: Patient's state
          example: NY
        zipCode:
          type: string
          nullable: true
          description: Patient's zip code
          example: '10001'
        bloodGroup:
          type: string
          nullable: true
          enum: ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-']
          description: Patient's blood group
          example: 'A+'
        emergencyContactName:
          type: string
          nullable: true
          description: Emergency contact person's name
          example: Jane Doe
        emergencyContactPhone:
          type: string
          nullable: true
          description: Emergency contact phone number
          example: '+1-555-0102'
        medicalHistory:
          type: string
          nullable: true
          description: Patient's medical history
          example: Hypertension diagnosed in 2020
        allergies:
          type: string
          nullable: true
          description: Patient's known allergies
          example: Penicillin
        isActive:
          type: boolean
          description: Whether the patient record is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the record was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the record was last updated

    CreatePatientDto:
      type: object
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
      properties:
        firstName:
          type: string
          maxLength: 100
          description: Patient's first name
          example: John
        lastName:
          type: string
          maxLength: 100
          description: Patient's last name
          example: Doe
        dateOfBirth:
          type: string
          format: date
          description: Patient's date of birth
          example: '1985-03-15'
        gender:
          type: string
          enum: [Male, Female, Other]
          description: Patient's gender
          example: Male
        email:
          type: string
          format: email
          nullable: true
          description: Patient's email address
          example: john.doe@email.com
        phoneNumber:
          type: string
          maxLength: 20
          description: Patient's phone number
          example: '+1-555-0101'
        address:
          type: string
          maxLength: 500
          nullable: true
          description: Patient's street address
        city:
          type: string
          maxLength: 100
          nullable: true
          description: Patient's city
        state:
          type: string
          maxLength: 100
          nullable: true
          description: Patient's state
        zipCode:
          type: string
          maxLength: 10
          nullable: true
          description: Patient's zip code
        bloodGroup:
          type: string
          nullable: true
          enum: ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-']
          description: Patient's blood group
        emergencyContactName:
          type: string
          maxLength: 200
          nullable: true
          description: Emergency contact person's name
        emergencyContactPhone:
          type: string
          maxLength: 20
          nullable: true
          description: Emergency contact phone number
        medicalHistory:
          type: string
          nullable: true
          description: Patient's medical history
        allergies:
          type: string
          nullable: true
          description: Patient's known allergies

    UpdatePatientDto:
      type: object
      required:
        - id
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phoneNumber
      properties:
        id:
          type: integer
          format: int32
          description: Patient's unique identifier
        firstName:
          type: string
          maxLength: 100
          description: Patient's first name
        lastName:
          type: string
          maxLength: 100
          description: Patient's last name
        dateOfBirth:
          type: string
          format: date
          description: Patient's date of birth
        gender:
          type: string
          enum: [Male, Female, Other]
          description: Patient's gender
        email:
          type: string
          format: email
          nullable: true
          description: Patient's email address
        phoneNumber:
          type: string
          maxLength: 20
          description: Patient's phone number
        address:
          type: string
          maxLength: 500
          nullable: true
          description: Patient's street address
        city:
          type: string
          maxLength: 100
          nullable: true
          description: Patient's city
        state:
          type: string
          maxLength: 100
          nullable: true
          description: Patient's state
        zipCode:
          type: string
          maxLength: 10
          nullable: true
          description: Patient's zip code
        bloodGroup:
          type: string
          nullable: true
          enum: ['A+', 'A-', 'B+', 'B-', 'AB+', 'AB-', 'O+', 'O-']
          description: Patient's blood group
        emergencyContactName:
          type: string
          maxLength: 200
          nullable: true
          description: Emergency contact person's name
        emergencyContactPhone:
          type: string
          maxLength: 20
          nullable: true
          description: Emergency contact phone number
        medicalHistory:
          type: string
          nullable: true
          description: Patient's medical history
        allergies:
          type: string
          nullable: true
          description: Patient's known allergies

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message

    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string